Bustage fix for Ubuntu packaging fun.
More fun with M2Crypto. OOTB, the C code parts of M2Crypto won't build
on Ubuntu when installing the package.

The normal Ubuntu method of installing "python-m2crypto" only installs
for the system library. That doesn't work for all the virtual
environments used for testing & development.

Detailed notes are in tox.ini, as that is where more virtualenvs are
created. Pointer to that in the main Makefile