Rework Python package dependencies
The prior dependencies were not properly pinned, and some
upgrades-over-time caused breakage. (E.g. pyasn1 -- 0.4 is current, but
we don't work with it & need to stick with 0.2.3)

Also, the most consistent results with M2Crypto are to compile from
source. To inform pip of this preference, we need to specify the repo's
URL.

These changes also include the packages needed by the mar code. And some
housekeeping (repo location).