Update tests to support new Cert
Bug 1554767 is activating a new windows signing certificate. Add files
signed with the new key for testing.

Note that, since last work on the project, AWS is enforcing clock skew
checks on the APIs. Event times are now required to be within 15 minutes
of current server time. Updated the generation of actual lambda
invocation payloads to set a reasonable event time.